requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor>=3.7
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.51.0
sentence-transformers==3.1.1
pymongo>=4.11
httpx==0.27.2
